pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Website Analyzer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "ruff>=0.12.7",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "black>=25.1.0",
    "psycopg2>=2.9.10",
    "bs4>=0.0.2",
    "requests>=2.32.4",
    "validators>=0.35.0",
    "bootstrap-flask>=2.5.0",
]

